requirements.txt
pygame-ce>=2.5.0
pytmx>=3.31

